requirements.txt
colorlog==6.8.2
ensure==1.0.4
joblib==1.3.2
lightgbm==4.3.0
numpy==1.26.4
pandas==2.2.1
python_box==7.1.1
PyYAML==6.0.1
scikit_learn==1.4.0
setuptools==68.2.2
torch==2.2.1
torchvision==0.17.1
wfdb==4.1.2
xgboost==2.0.3
notebook
matplotlib
python-box
pathlib
tqdm
types-pyYAML
-e .
